+++ Package.swift
@@ -8,7 +8,7 @@ let package = Package(
     products: [.library(name: "GyoshukuKit", targets: ["GyoshukuKit"])],
     dependencies: [.package(path: "../KaitoKit")],
     targets: [
-        .target(name: "GyoshukuKit", linkerSettings: [.linkedLibrary("z")]),
+        .target(name: "GyoshukuKit", dependencies: [.product(name: "KaitoKit", package: "KaitoKit")], linkerSettings: [.linkedLibrary("z")]),
         .testTarget(
             name: "GyoshukuKitTests",
             dependencies: ["GyoshukuKit", .product(name: "KaitoKit", package: "KaitoKit")]